changed drivetime datatype from int to long

+++ dotnet/corecomponents/net.atos.daf.ct2.reportschedular/entity/TripDetails.cs
@@ -30,7 +30,7 @@ namespace net.atos.daf.ct2.reportscheduler.entity
 
         public double FuelConsumed { get; set; }
 
-        public int DrivingTime { get; set; }
+        public long DrivingTime { get; set; }
 
         public int Alert { get; set; }
